[packages] php5: Add support for using the system timezone database
Using the systems database rather than embedding a copy has several
advantages. Its easier to keep the database up-to-date and it can
be used by multiple applications. So some bytes in the firmware image
can be saved, especially when only required timezones are installed.

Additionally, it not necessary anymore to configure the timezone
explicitly via php.ini (required since PHP 5.3). Instead OpenWRTs
system-wide settings are used.

This new approach is enabled by default, however it is possible
to deselect this feature to restore PHP's original behaviour.

The patches are taken from the Debian PHP patches.

Signed-off-by: Peter Naulls <peter@chocky.org>
Signed-off-by: Michael Heimpold <mhei@heimpold.de>