libjpeg-turbo: add package for test utils, move appropriate targets there
Trying to figure out where jpeggut came from, found a) that it was a
typo and b) that it was renamed in r643 from jpegut to tjunittest.  In
the same revision, jpegtest became tjbench, and later in r733, jcstest
was added to test jpeg color space extentions.  Since these tools will
only be desirable for testing, added a new package libjpeg-turbo-tests
to provide these utilities.  The utilities needed libturbojpeg, so
added what had been a commented out library to the new package.

The libturbojpeg are currently commented out in InstallDev as well.
If external packages begin to rely on libturbojpeg functions for some
reason, the library should be added back.

Signed-off-by: Russell Senior <russell@personaltelco.net>
Signed-off-by: Florian Fainelli <florian@openwrt.org>