mosquitto: bump 1.2 to 1.2.3
This includes UCI support for a new config option, bridge_tls_version.
Add bridge tls vesion support

Updates mosquitto, the MQTT broker, and client libraries to version
1.2.2, from version 1.2.0.  Changes are bugfixes, most relevant to
OpenWrt being that the -nossl package variant actually builds again,
and greatly improved SSL performance.

This includes the full changes for 1.2.3:
http://mosquitto.org/2013/12/version-1-2-3-released/

This includes the full changes for 1.2.2:
http://mosquitto.org/2013/10/version-1-2-2-released/

This includes the full changes for 1.2.1:
http://mosquitto.org/2013/09/version-1-2-1-released/

Signed-off-by: Karl Palsson <karlp@remake.is>