lantiq: BTHOMEHUBV2B use bigger mtd partition for kernel
The bb-rc3 image for the BTHOMEHUBV2B is too big for its
mtd partition. This patch corrects the partition sizes in
the device tree. This patch should really go in before
bb-final, otherwise the BTHOMEHUBV2B images won't be useable.
I do apologise for not spotting this straight away.

Many thanks,

Ben

Signed-off-by: Ben Mulvihill <ben.mulvihill@gmail.com>

Backport of r42316