include/scan.mk: Add -a argument to the grep to treat all Makefiles as a text file.
Hello,

I found out that in some rare cases grep can treat Makefile as a binary file. That happened to me on UTF-8 Gentoo if Makefile contained a character which was from extended ASCII table.
Without this patch the output of the grep command in the $(FILELIST) target is not valid (contains line "Binary file matches") and following commands fail to create feed index file.

Best Regards,
Martin Strbacka

Signed-off-by: Martin Strbacka <martin.strbacka@nic.cz>