ar71xx: rb493g: fix gpio-mdio accesses under load
It seems that there are maximum timings for mdio accesses that can be
hit when the system is under load and the thread is scheduled during
a read or write access. Since there is no way of knowing if this
happens as there is not even a parity bit, try to work around it
by disabling interrupts during any gpio-mdio accesses.

Performance impact seems to be neglectable, as mdio accesses are not
that often.

Hopefully fixes #19500.

Signed-off-by: Jonas Gorski <jogo@openwrt.org>

Backport of r46012.