mac80211: use wl18xx-fw-4.bin
The driver only supports wl18xx-fw-4.bin and does not support the old
firmware wl18xx-fw-3.bin and more.

The wl18xx-conf.bin file was removed upstream it does not fit all
devices any more, see this on how to generate one:
http://processors.wiki.ti.com/index.php/Open_Source_Wireless_Connectivity_wlconf

If someone has a config file for the common platforms with the wifi nic
OpenWrt uses please send a patch.

Signed-off-by: Hauke Mehrtens <hauke@hauke-m.de>