malta: set CPU_TYPE:=24kc.
There are instructions (e.g. wsbh) only available in MIPS32R2 which is
the isa level used by many MIPS targets in OpenWrt.  As we use the malta
target mostly for development and testing purposes, setting the CPU_TYPE
to 24Kc will make the emulation more real.

The following is output from qemu-system-mipsel 2.3.0

    root@OpenWrt:/# cat /proc/cpuinfo  | grep 'cpu model\|isa'
    cpu model               : MIPS 24Kc V0.0  FPU V0.0
    isa                     : mips1 mips2 mips32r1 mips32r2

Signed-off-by: Yousong Zhou <yszhou4tech@gmail.com>